+++ _universum/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: UTF-8 -*-
 
 __title__ = "Universum"
-__version__ = "0.14.0"
+__version__ = "0.14.1"