+++ _universum/module_arguments.py
@@ -21,10 +21,10 @@ class ModuleNamespace(argparse.Namespace):
             try:
                 ns = self.__dict__[group]
             except KeyError:
-                raise AttributeError
+                raise AttributeError("No attribute '" + name + "' in module arguments")
             return getattr(ns, name)
         else:
-            raise AttributeError
+            raise AttributeError("No attribute '" + name + "' in module arguments")
 
 
 # noinspection PyProtectedMember